libbladeRF: Be explicit with RXVGA2 cal.
It was seen that the VCM will sometimes come back in a bad state.
This register has a very well known value that it should be.  Setting
it to be these explicit values seems to help alleviate putting the
LMS is a bad state.

Investigation is required to figure out why the VCM is coming back
scrogged.